+++ src/main/java/org/openpredict/exchange/beans/Order.java
@@ -19,7 +19,7 @@ public class Order extends OrderCommand {
     public long filled;
 
     @Builder(builderMethodName = "orderBuilder", builderClassName = "OrderBuilder")
-    public Order(OrderCommandType command, long orderId, int symbol, int price, long size, OrderAction action, OrderType orderType,
+    public Order(OrderCommandType command, long orderId, int symbol, long price, long size, OrderAction action, OrderType orderType,
                  long uid, long timestamp, long filled) {
         //super(command, orderId, symbol, price, size, action, orderType, uid, timestamp, 0, null, null);
         super(command, orderId, symbol, price, size, action, orderType, uid, timestamp, 0, null, null);
